# docs/openapi.yaml
openapi: 3.1.0
info:
  title: QUANT INDUSTRY V1 API
  description: |
    Institutional-grade quantitative trading platform API.
    
    ## Overview
    This API provides endpoints for:
    - Real-time market data and quotes
    - Trading signals and order management
    - ML/AI brain for signal generation
    - Risk management and monitoring
    - Portfolio tracking and analytics
    - Options flow and GEX analysis
    - Backtesting and strategy optimization
    
    ## Authentication
    Currently supports API key authentication via environment variables.
    Future: JWT-based auth for multi-tenant support.
    
    ## WebSocket
    Real-time data available via WebSocket at `/ws/unified`
    
  version: 1.0.0
  contact:
    name: QUANT INDUSTRY Support
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: http://localhost:8001
    description: Backend server (alternative)
...
paths:
...
  /api/market/tickers:
    get:
      tags: [market]
      summary: Get ticker prices
      operationId: getTickers
      parameters:
        - name: symbols
          in: query
          schema:
            type: string
            default: SPY,QQQ,DIA,IWM
          description: Comma-separated list of symbols
      responses:
        '200':
          description: Ticker prices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TickerPrice'
  /api/market/quote/{symbol}:
    get:
      tags: [market]
      summary: Get detailed quote
      operationId: getQuote
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Quote data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
...
  /api/options/chain/{symbol}:
    get:
      tags: [options]
      summary: Get options chain
      operationId: getOptionsChain
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
        - name: expiration
          in: query
          schema:
            type: string
        - name: option_type
          in: query
          schema:
            type: string
            enum: [ALL, CALL, PUT]
            default: ALL
      responses:
        '200':
          description: Options chain
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OptionContract'
  /api/gex/{symbol}:
    get:
      tags: [options]
      summary: Get gamma exposure
      operationId: getGEX
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: GEX data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GEXData'
...
components:
  schemas:
...
    TickerPrice:
      type: object
      properties:
        symbol:
          type: string
        price:
          type: number
        change:
          type: number
        change_pct:
          type: number
        volume:
          type: integer
    Quote:
      type: object
      properties:
        symbol:
          type: string
        price:
          type: number
        bid:
          type: number
        ask:
          type: number
        change:
          type: number
        change_pct:
          type: number
        volume:
          type: integer
        high:
          type: number
        low:
          type: number
        open:
          type: number
        prev_close:
          type: number
        source:
          type: string
        timestamp:
          type: string
          format: date-time
...
    OptionContract:
      type: object
      properties:
        expiration:
          type: string
        strike:
          type: number
        option_type:
          type: string
          enum: [call, put]
        bid:
          type: number
        ask:
          type: number
        iv:
          type: number
        delta:
          type: number
        gamma:
          type: number
        theta:
          type: number
        vega:
          type: number
        oi:
          type: integer
        volume:
          type: integer
    GEXData:
      type: object
      properties:
        symbol:
          type: string
        total_gex:
          type: number
        call_gex:
          type: number
        put_gex:
          type: number
        key_levels:
          type: array
          items:
            type: object
            properties:
              strike:
                type: number
              gex:
                type: number
        flip_point:
          type: number
...